+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - refactor: improve type stubs for kivy app by only providing stubs for pyx files and making `Widget` use correct `EventListener` base and fix type errors coming to surface
 - refactor(core): make `get_eeprom_data` return a default value instead of `None` if there is an issue in procuring the data
 - refactor(docker): remove sync wrapper functions around async actions and make the action directly async as it is run by `subscribe_event` and it supports async functions, general housekeeping for docker service
+- fix: separate `has_output` and `has_output_stream` to fix the old `send_command` callers being broken with the new `send_command` implementation
 
 ## Version 1.3.0